instance: fix issue with chosing the instance
The return type has been changed from *string to string, not only
because that solves the issue, but because it did not make that much
sense to have a pointer as it was always dereferenced in the calling
method. The error is also now verified in the calling method.

Finally couple logging call have been 'fixed' (removed new line) in
order to avoid messing with the overall output.